decouple participantcontext service

+++ core/identity-hub-participants/src/main/java/org/eclipse/edc/identityhub/participantcontext/ParticipantContextServiceImpl.java
@@ -16,9 +16,6 @@ package org.eclipse.edc.identityhub.participantcontext;
 
 import com.nimbusds.jose.jwk.JWK;
 import com.nimbusds.jose.jwk.JWKParameterNames;
-import org.eclipse.edc.iam.did.spi.document.DidDocument;
-import org.eclipse.edc.iam.did.spi.document.VerificationMethod;
-import org.eclipse.edc.identithub.did.spi.DidDocumentService;
 import org.eclipse.edc.identityhub.security.KeyPairGenerator;
 import org.eclipse.edc.identityhub.spi.ParticipantContextService;
 import org.eclipse.edc.identityhub.spi.events.ParticipantContextObservable;
@@ -38,11 +35,9 @@ import org.eclipse.edc.transaction.spi.TransactionContext;
 import java.security.KeyPair;
 import java.security.PublicKey;
 import java.text.ParseException;
-import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 
-import static java.util.Optional.ofNullable;
 import static org.eclipse.edc.spi.result.ServiceResult.badRequest;
 import static org.eclipse.edc.spi.result.ServiceResult.conflict;
 import static org.eclipse.edc.spi.result.ServiceResult.fromFailure;
@@ -63,18 +58,16 @@ public class ParticipantContextServiceImpl implements ParticipantContextService 
     private final TransactionContext transactionContext;
     private final ApiTokenGenerator tokenGenerator;
     private final KeyParserRegistry keyParserRegistry;
-    private final DidDocumentService didDocumentService;
     private final ParticipantContextObservable observable;
 
     public ParticipantContextServiceImpl(ParticipantContextStore participantContextStore, Vault vault, TransactionContext transactionContext,
-                                         KeyParserRegistry registry, DidDocumentService didDocumentService, ParticipantContextObservable observable) {
+                                         KeyParserRegistry registry, ParticipantContextObservable observable) {
         this.participantContextStore = participantContextStore;
         this.vault = vault;
         this.transactionContext = transactionContext;
         this.observable = observable;
         this.tokenGenerator = new ApiTokenGenerator();
         this.keyParserRegistry = registry;
-        this.didDocumentService = didDocumentService;
     }
 
     @Override
@@ -83,10 +76,8 @@ public class ParticipantContextServiceImpl implements ParticipantContextService 
             var apiKey = new AtomicReference<String>();
             var context = convert(manifest);
             var res = createParticipantContext(context)
-                    .compose(u -> generateAndStoreToken(context)).onSuccess(apiKey::set)
-                    .compose(ak -> createOrUpdateKey(manifest.getKey()))
-                    .compose(jwk -> generateDidDocument(manifest, jwk))
-                    .onSuccess(u -> observable.invokeForEach(l -> l.created(context)));
+                    .compose(u -> createTokenAndStoreInVault(context)).onSuccess(apiKey::set)
+                    .onSuccess(apiToken -> observable.invokeForEach(l -> l.created(context, manifest)));
             return res.map(u -> apiKey.get());
         });
     }
@@ -98,7 +89,7 @@ public class ParticipantContextServiceImpl implements ParticipantContextService 
             if (res.succeeded()) {
                 return res.getContent().stream().findFirst()
                         .map(ServiceResult::success)
-                        .orElse(notFound("No ParticipantContext with ID '%s' was found.".formatted(participantId)));
+                        .orElse(notFound("ParticipantContext with ID '%s' does not exist.".formatted(participantId)));
             }
             return fromFailure(res);
         });
@@ -107,14 +98,18 @@ public class ParticipantContextServiceImpl implements ParticipantContextService 
     @Override
     public ServiceResult<Void> deleteParticipantContext(String participantId) {
         return transactionContext.execute(() -> {
-            var did = ofNullable(findByIdInternal(participantId)).map(ParticipantContext::getDid);
+            var participantContext = findByIdInternal(participantId);
+            if (participantContext == null) {
+                return ServiceResult.notFound("A ParticipantContext with ID '%s' does not exist.");
+            }
+
             var res = participantContextStore.deleteById(participantId);
             if (res.failed()) {
                 return fromFailure(res);
             }
 
-            return did.map(d -> didDocumentService.unpublish(d).compose(u -> didDocumentService.deleteById(d)))
-                    .orElseGet(ServiceResult::success);
+            observable.invokeForEach(l -> l.deleted(participantContext));
+            return ServiceResult.success();
         });
     }
 
@@ -125,7 +120,7 @@ public class ParticipantContextServiceImpl implements ParticipantContextService 
             if (participantContext.failed()) {
                 return participantContext.map(pc -> null);
             }
-            return generateAndStoreToken(participantContext.getContent());
+            return createTokenAndStoreInVault(participantContext.getContent());
         });
     }
 
@@ -144,24 +139,12 @@ public class ParticipantContextServiceImpl implements ParticipantContextService 
 
     }
 
-    private ServiceResult<String> generateAndStoreToken(ParticipantContext participantContext) {
+    private ServiceResult<String> createTokenAndStoreInVault(ParticipantContext participantContext) {
         var alias = participantContext.getApiTokenAlias();
         var newToken = tokenGenerator.generate(participantContext.getParticipantId());
         return vault.storeSecret(alias, newToken)
                 .map(unused -> success(newToken))
                 .orElse(f -> conflict("Could not store new API token: %s.".formatted(f.getFailureDetail())));
-    }
-
-    private ServiceResult<Void> generateDidDocument(ParticipantManifest manifest, JWK publicKey) {
-        var doc = DidDocument.Builder.newInstance()
-                .id(manifest.getDid())
-                .service(manifest.getServiceEndpoints().stream().toList())
-                .verificationMethod(List.of(VerificationMethod.Builder.newInstance()
-                        .publicKeyJwk(publicKey.toJSONObject())
-                        .build()))
-                .build();
-        return didDocumentService.store(doc, manifest.getParticipantId())
-                .compose(u -> manifest.isActive() ? didDocumentService.publish(doc.getId()) : success());
     }
 
     private ServiceResult<JWK> createOrUpdateKey(KeyDescriptor key) {
